Rename `routes.dart` to `app_config.dart` and moved to the `config` package
