--- lib/config/api_config.dart
@@ -1,4 +1,0 @@
-class ApiConfig {
-  static const String baseUrl = 'our api url';
-  // Add more API endpoints here
-}
