+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gocast_mobile/base/networking/api/handler/grpc_handler.dart';
+import 'package:gocast_mobile/config/app_config.dart';
 import 'package:gocast_mobile/models/error/error_model.dart';
 import 'package:gocast_mobile/models/user/user_state_model.dart';
 import 'package:gocast_mobile/view_models/UserViewModel.dart';
@@ -8,7 +9,6 @@ import 'package:gocast_mobile/views/course_view/courses_overview_view.dart';
 import 'package:gocast_mobile/views/login_view/internal_login_view.dart';
 import 'package:gocast_mobile/views/on_boarding_view/welcome_screen_view.dart';
 import 'package:gocast_mobile/views/utils/globals.dart';
-import 'package:gocast_mobile/views/utils/routes.dart';
 import 'package:gocast_mobile/views/utils/theme.dart';
 import 'package:logger/logger.dart';
